labeling: refactor tests a bit
